+++ packages/core/e-charts/src/index.ts
@@ -1,17 +1,13 @@
-// import type { App } from 'vue'
 import ECharts from './components/ECharts.vue'
+import HeatmapChart from './components/chart-types/HeatmapChart.vue'
+import TreeMapChart from './components/chart-types/TreeMapChart.vue'
 
-// Export Vue plugin
-// We rarely want to export components as a plugin as we prefer to support proper tree-shaking in the host application. Only enable if you're packing a Vue plugin.
-// export default {
-//   // Customize Vue plugin options as desired
-//   // Providing a `name` property allows for customizing the registered
-//   // name of your component (useful if exporting a single component).
-//   install: (app: App, options: { name?: string, [key: string]: any } = {}): void => {
-//     app.component(options.name || 'ECharts', ECharts)
-//   },
-// }
+export {
+  ECharts, // Base ECharts component, exported by wrapping up with default theme and common settings
+  HeatmapChart,
+  TreeMapChart,
+}
 
-export { ECharts }
+export { chartColors, useChartColors } from './composables/useChartColors.ts'
 
-export * from './types'
+export * from './types/index.ts'