+++ packages/core/e-charts/CHANGELOG.md
@@ -3,6 +3,35 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.1](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fe-charts%400.4.0...%40kong-ui-public%2Fe-charts%400.4.1) (2026-09-24)
+
+**Note:** Version bump only for package @kong-ui-public/e-charts
+
+
+
+
+
+# [0.4.0](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fe-charts%400.3.0...%40kong-ui-public%2Fe-charts%400.4.0) (2026-09-24)
+
+### Features
+
+* **e-charts:** add a shared tooltip matching analytics-chart ([#3871](https://github.com/Kong/public-ui-components/issues/3871)) ([51dda62](https://github.com/Kong/public-ui-components/commit/51dda62d9cfb9a0c4d5ed2630cbde8a0d475ac61))
+
+
+# [0.3.0](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fe-charts%400.2.0...%40kong-ui-public%2Fe-charts%400.3.0) (2026-09-24)
+
+### Features
+
+* **e-charts:** add chart theming and heatmap customization ([#3869](https://github.com/Kong/public-ui-components/issues/3869)) ([d2f8291](https://github.com/Kong/public-ui-components/commit/d2f8291350667115a188602ee8afa27b89108999))
+
+
+# [0.2.0](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fe-charts%400.1.1...%40kong-ui-public%2Fe-charts%400.2.0) (2026-09-24)
+
+### Features
+
+* **echarts:** add support for heatmap [OLLY-1320] ([#3835](https://github.com/Kong/public-ui-components/issues/3835)) ([9a84c3b](https://github.com/Kong/public-ui-components/commit/9a84c3b3dce06f78b1d7885724c4e76607ac0d18))
+
+
 ## [0.1.1](https://github.com/Kong/public-ui-components/compare/%40kong-ui-public%2Fe-charts%400.1.0...%40kong-ui-public%2Fe-charts%400.1.1) (2026-09-21)
 
 **Note:** Version bump only for package @kong-ui-public/e-charts